Po dalším vyčištění kódu všech managerů.

+++ src/com/certifikace/projekt1/ActualMenuManager.java
@@ -18,24 +18,9 @@ public class ActualMenuManager {
 
     public ActualMenuManager() {this.amList = new ArrayList<>();}
 
-
-    private boolean firstWriteDetector(ActualMenu actualMenu) {
-        return actualMenu.getAmNumberOfNextCategories() == 0 && actualMenu.getAmTitle().equals("Empty Title")
-                && actualMenu.getAmQuantity() == 0  && actualMenu.getAmPrice().compareTo(BigDecimal.ZERO) == 0
-                && actualMenu.getAmPreparationTime() == 999999 && actualMenu.getAmNumberOfNextPhotos() == 0;
-    }
-    private void removefirstWrite() {
-        Iterator<ActualMenu> iterator = amList.iterator();
-        while (iterator.hasNext()) {
-            ActualMenu actualMenu = iterator.next();
-            if (firstWriteDetector(actualMenu)) {iterator.remove();}
-        }
-    }
     public void addFoodToMenu(String title, int quantity, DishManager dishManager) throws RestaurantException {
-        // Když tam bude první programem vytvořený zápis po prvním spuštěnmí, odstraní se z Listu
-        removefirstWrite();
-
         String helpSameErrMessage = " Jídlo NEBYLO přidáno do amList!";
+
         for (Dish dish : dishManager.getDishList()) {
             if (dish.dishDetectSameTitleAndQuantity(title, quantity)) {
                 for (ActualMenu existingFood : amList) {
@@ -82,84 +67,10 @@ public class ActualMenuManager {
         }
     }
 
-    public void clearAmList() throws IOException {
-        createBackUpMenuToFile(); amList.clear(); createEmptyDishsFile(fileActualMenu());
-    }
-
-    private void createEmptyDishsFile(String fileActualMenu) {
-        try (BufferedWriter writer = new BufferedWriter(new FileWriter(fileActualMenu))) {
-            writer.write(delimiter() + 0 + delimiter() + "Empty Title" + delimiter() + 0 + delimiter() + delimiter()
-                    + "0" + delimiter() + 999999 + delimiter() + delimiter() + 0); writer.newLine();
-        } catch (IOException e) {
-            System.err.println("Chyba při vytváření souboru při neexistenci souboru s aktuálním menu: "
-                    + e.getMessage());
-        }
-    }
-
-
-
-    /*
-    public void loadDataMenuFromFile(String fileActualMenu, String delimiter) throws RestaurantException {
-        if (!Files.exists(Paths.get(fileActualMenu))) {createEmptyDishsFile(fileActualMenu); return;}
-        int i; int helpBadFormatIdentificator = 0; FoodCategory helpCategory = null;
-        String line = ""; String[] item = new String[0];
-        FoodCategory amMainCategory; int amNumberOfNextCategories; String amTitle; int amQuantity;
-        String amUnitOfQuantity; BigDecimal amPrice; int amPreparationTime; String amMainPhoto;
-        int amNumberOfNextPhotos;
-        try (Scanner scannerLoadData = new Scanner(new BufferedReader(new FileReader(fileActualMenu)))) {
-            while (scannerLoadData.hasNextLine()) {
-                line = scannerLoadData.nextLine();
-                item = line.split(delimiter);
-                amMainCategory = FoodCategory.valueOf(item[0]);
-                if (amMainCategory == null) {System.err.println("Chyba: V souboru DB-ActualMenu.txt je "
-                        + "základní kategorie, která má hodnotu null nebo má neplatný formát nebo neexistuje"
-                        + " v seznamu kategorií ve FoodCategory na řádku s obsahem: " + line);}
-                helpBadFormatIdentificator = 1;
-                amNumberOfNextCategories = Integer.parseInt(item[1]);
-                List<FoodCategory> amNextCategory = new ArrayList<>();
-                for (i = 0; i < amNumberOfNextCategories; i++) {
-                    helpCategory = FoodCategory.valueOf(item[i + 2]);
-                    amNextCategory.add(helpCategory);
-                    if (helpCategory == null) {System.err.println("Chyba: V souboru DB-ActualMenu.txt je další "
-                            + "kategorie, která má hodnotu null nebo má neplatný formát nebo neexistuje v seznamu "
-                            + "kategorií ve FoodCategory na řádku s obsahem: " + line);}
-                }
-                amTitle = item[2 + amNumberOfNextCategories];
-                helpBadFormatIdentificator = 2 + amNumberOfNextCategories;
-                amQuantity = Integer.parseInt(item[3 + amNumberOfNextCategories]);
-                amUnitOfQuantity = item[4 + amNumberOfNextCategories];
-                helpBadFormatIdentificator = 4 + amNumberOfNextCategories;
-                amPrice = new BigDecimal(item[5 + amNumberOfNextCategories]);
-                helpBadFormatIdentificator = 5 + amNumberOfNextCategories;
-                amPreparationTime = Integer.parseInt(item[6 + amNumberOfNextCategories]);
-                amMainPhoto = item[7 + amNumberOfNextCategories];
-                helpBadFormatIdentificator = 7 + amNumberOfNextCategories;
-                amNumberOfNextPhotos = Integer.parseInt(item[8 + amNumberOfNextCategories]);
-                if (item.length != 9 + amNumberOfNextCategories + amNumberOfNextPhotos) {
-                    throw new RestaurantException("Chyba: Špatný počet položek na řádku: " + line);
-                }
-                List<String> amNextPhoto = new ArrayList<>();
-                for (i = 0; i < amNumberOfNextPhotos; i++) {
-                    amNextPhoto.add(item[i + 9 + amNumberOfNextCategories]);
-                }
-                ActualMenu newActualMenu = new ActualMenu(amMainCategory, amNumberOfNextCategories, amNextCategory,
-                        amTitle, amQuantity, amUnitOfQuantity, amPrice, amPreparationTime, amMainPhoto,
-                        amNumberOfNextPhotos, amNextPhoto);
-                amList.add(newActualMenu);
-            }
-        } catch (FileNotFoundException e) {
-            throw new RestaurantException("Chyba: Soubor " + fileActualMenu + " nebyl nalezen! "
-                    + e.getLocalizedMessage());
-        } catch (NumberFormatException e) {
-            throw new RestaurantException("Chyba: V souboru DB-ActualMenu.txt není číslo nebo má nedovolenou zápornou "
-                    + "hodnotu na řádku: " + line + " položka č." + helpBadFormatIdentificator);
-        }
-    }
-
-     */
+    public void clearAmList() {amList.clear();}
 
     public void loadDataMenuFromFile(String fileActualMenu, String delimiter) throws RestaurantException {
-        if (!Files.exists(Paths.get(fileActualMenu))) { createEmptyDishsFile(fileActualMenu); return; }
+        if (!Files.exists(Paths.get(fileActualMenu))) {return;}
         String helpSameErrMessage1 = "Chyba: Špatný počet položek na řádku č. ";
         String helpSameErrMessage2 = " Vadná položka je na řádku v pořadí ";
         String helpSameErrMessage3 = " NEBO! Na tomto řádku je špatný počet položek díky proměnlivému počtu dalších "
@@ -225,16 +136,12 @@ public class ActualMenuManager {
         }
     }
 
-    public void createBackUpMenuToFile() throws IOException {
-        File originalFile = new File(fileActualMenu()); File backupFile = new File(fileBackUpMenu());
-        Files.copy(originalFile.toPath(), backupFile.toPath(), StandardCopyOption.REPLACE_EXISTING);
-    }
     public void saveDataMenuToFile(String fileName) throws RestaurantException {
         try {
-            File originalFile = new File(fileActualMenu());
-            //File backupFile = new File(fileBackUpMenu());
-            //Files.copy(originalFile.toPath(), backupFile.toPath(), StandardCopyOption.REPLACE_EXISTING);
-            createBackUpMenuToFile();
+            File originalFile = new File(fileActualMenu()); File backupFile = new File(fileBackUpMenu());
+            if (Files.exists(Paths.get(originalFile.toURI()))) {
+                Files.copy(originalFile.toPath(), backupFile.toPath(), StandardCopyOption.REPLACE_EXISTING);
+            }
             try (BufferedWriter writer = new BufferedWriter(new FileWriter(fileName))) {
                 for (ActualMenu actualMenu : amList) {
                     writer.write(actualMenu.getAmMainCategory() + delimiter());
